+++ server/cmd/localserver/main.go
@@ -5,24 +5,22 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/mleone10/expense-system/adapters/auth"
+	"github.com/mleone10/expense-system/adapters/googleauth"
 	"github.com/mleone10/expense-system/adapters/rest"
 	"github.com/mleone10/expense-system/adapters/stdlogger"
 )
 
 func main() {
-	authClient := auth.NewAuthClient(
-		auth.WithClientHostname("localhost:3000"),
-		auth.WithClientScheme("http"),
-		auth.WithCognitoClientId("6ka3m790cv5hrhjdqt2ju89v45"),
-		auth.WithCognitoClientSecret(os.Getenv("COGNITO_CLIENT_SECRET")),
+	authClient := googleauth.NewAuthClient(
+		googleauth.WithClientHostname("localhost:3000"),
+		googleauth.WithClientScheme("http"),
+		googleauth.WithCognitoClientId("6ka3m790cv5hrhjdqt2ju89v45"),
+		googleauth.WithCognitoClientSecret(os.Getenv("COGNITO_CLIENT_SECRET")),
 	)
-
-	logger := stdlogger.Logger{}
 
 	server, _ := rest.NewServer(
 		rest.WithAuthClient(authClient),
-		rest.WithLogger(logger),
+		rest.WithLogger(stdlogger.Logger{}),
 	)
 
 	fmt.Println(http.ListenAndServe(":8080", server))
